\name{mendota}
\alias{mendota}
\docType{data}
\title{Number of freezing days at Lake Mendota}
\description{This dataset shows the number of freezing days at Lake Mendota measured from November, 23, in the year 1854. 
}
\usage{data(mendota)}
\format{A data frame with 12 subjects.
}
\references{
Bhattacharyya, G. K., & Klotz, J. H. (1966). The bivariate trend of Lake Mendota. Technical Report No. 98, Department of Statistics, University of Wisconsin.

Barlow, R. E., Bartholomew, D. J., Bremner, J. M., & Brunk, H. D. (1972). Statistical inference under order restrictions: The theory and application of isotonic regression. Chichester: Wiley.
}
\examples{
data(mendota)
}
\keyword{datasets}
